requirements.txt
python-telegram-bot==20.7
openai==1.3.7
schedule==1.2.0
python-dotenv==1.0.0
asyncio
